Remove podcast string from move_file.py
podcast was only used to determine the complete archive path, but it's cleaner to
have the caller compute the archive path and just pass it indirectly.

Also change destination to file_destination and pass in the full path to be consistent with
how the file and archive location are passed in.