Move ffmpeg specific code into conversions.py (#22)
conversion.py is the only spot that requires ffmpeg interactions, so we
moved the one function there and deleted ffmpeg_helper.py

The end result is a cleaner interface, where the implementation detail
(that we use ffmpeg) is hidden in a single file without making that file
more complex.

Also remove file deletion to inside _ConvertFile, as Convert suggests
the original file doesn't exists afterwards.