Renamed AndroidPhone.connected_to_phone to AndroidPhone.connect_to_phone
To try and make it clearer that this isn't a one time check, but actively trying to connect with retries possible.

Also slightly rewrote connect_to_phone to be a bit clearer to read and some changes to the related tests, as there were two different groups of tests with some repeated tests.